moved image styles to #wiki_page_show
Previously all image style were nested under #content. I added a 10px
margin to the top of all images to make them flush with paragraphs in
adjacent columns. Being under #content affected images outside of the
page container (such as the icons in the content editor).

This change makes sure that only images in the template are being
effected.